moved fetch landmarks to helper functions
